Query params can now be passed for config
